+++ client/__init__.lua
@@ -1,8 +1,7 @@
 class 'XP4FD_CQBTClient'
 
 function XP4FD_CQBTClient:__init()
-    print('Hello world!')
-
+    print('Hello client!')
 end
 
-g_XP4FD_CQBTClient = XP4FD_CQBTClient();
+g_XP4FD_CQBTClient = XP4FD_CQBTClient()