+++ shared/__init__.lua
@@ -1,8 +1,1 @@
-class 'XP4FD_CQBTShared'
-
-function XP4FD_CQBTShared:__init()
-    print('Hello world!')
-
-end
-
-g_XP4FD_CQBTShared = XP4FD_CQBTShared();
+require('__shared/util/registerSectors')