+++ ChangeLog.md
@@ -1,2 +1,3 @@
 # Changes per commit
-11/11/20 - Added US sector 1 Map vector Data and Flag data from CQ-L
+- 11/11/20 - Added US sector 1 Map vector Data and Flag data from CQ-L
+- 11/12/20 - Seperated the logic in various files and updated GUID for TDM